Ana/3699 refactor to delete components/wallet and add session folder (#5068)
* delete components/wallet folder

* move also unit tests files

* move all the session components to session folder

* move also unit tests session

* change the folder in webpack

* fix session unit tests location

* move also TmLiSession to session folder

* changelog

* changelog

* changelog
