+++ app/changes/ana_3699-refactor-to-delete-components_wallet
@@ -1,0 +1,2 @@
+[Repository] [#3699](https://github.com/cosmos/lunie/issues/3699) Delete the components/wallet folder @Bitcoinera
+[Repository] [#5068](https://github.com/cosmos/lunie/pull/5068) Add the components/session folder @Bitcoinera
