+++ app/src/components/session/TmSessionExplore.vue
