+++ app/src/components/session/TmSessionHardware.vue
