+++ extension/changes/ana_3699-refactor-to-delete-components_wallet
@@ -1,0 +1,1 @@
+[Repository] [#5068](https://github.com/cosmos/lunie/pull/5068) Add the components/session folder @Bitcoinera
